login+bugfixes
Users now can properly register, log in, and stay loged in as long as the browser is not closed or they do not press the 'log out' button.
They also get redirected to the proper page if they try to reenter to the main page and some other places that should not happen.
There were some minor bugfixes with said redirects as well.